Rename providers/tournament.ts to controllers/tournamentController.ts.
